requirements.txt
Flask==3.1.3
Adafruit_DHT==1.4.0
RPi.GPIO==0.7.1